MB-45185: Pass CompactionCallbacks to low level flush accounting [1/3]
For magma we need to stop counting deleted docs in collection item
sizes as we can't decrement the collection sizes correctly when we
purged tombstones as we call the compaction callback for old revisions.
We'll use the compactionCallbacks value in a future change to update
the collection disk size accordingly.

Change-Id: Iea19a4db401cde1857237300eedac432e95ad21d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166344
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>