MB-45185: Rename WantsDropped to CompactionCallbacks
To decrement the collection disk size by the full value rather than
the tombstone for ONLY magma we need to be able to determine between
couchstore and magma in FlushAccounting. We already pass in
WantsDropped which can be read as IsMagma. WantsDropped does not mean
anything in this new case. WantsDropped was originally added to deal
with collection item counting which differs between couchstore and
magma because magma has stale items and invokes the compaction
callback for them.

Rename WantsDropped to CompactionCallbacks with options for
AnyRevision and LatestRevision to better encapsulate the reasons
for the difference in behaviour.

Change-Id: I326b1c38b14b5895bbc81b2ad1ed78a965718f3f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166354
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>