MB-45185: [Magma] Decrement coll size by full value at deletion [3/3]
Currently we decrement collection sizes when deleting items in them
by the difference in the number of bytes that are/will be stored on
disk. An issue exists when purging tombstones though; it's not possible
to assert if we are looking at the most recent version of an item
when iterating items in the compaction callback, as such, we cannot
decrement the item count by the remaining bytes (of the tombstone)
as stale versions may exist.

Fix this issue by not tracking tombstones in the collection disk size
for magma.

Change-Id: Ibff066623a6920bc09d78e1481001237d4d3d3b6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166345
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>