TSAN: Ensure we don't try and access freed ThreadGate
TSAN issue has been observed in setupForImplicitCompactionTest() where
we try and access the ThreadGate tg from
postPurgeSeqnoImplicitCompactionHook() after it's gone out of scope,
due to setupForImplicitCompactionTest() finishing execution. This is
due to Magma triggering a second implicit compaction after the first
one fails.

To fix this, refactor the test to simulate an implicit compaction by
creating a new Magma checkpoint and then calling
MockMagmaKVStore::runImplicitCompactKVStore() to simulate an implicit
compaction being run. This allows us to ensure that there is only one
compaction run in the test.

Change-Id: Iaffe2c0dd0d96152164ba60b5654658e1a3e5114
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166423
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>