MB-47318: Address double acking when closing a stream
When the DcpConsumerTask used 'moveFromFront' and took ownership
of DcpResponse, the Buffer::bytes member still accounted for
the size of the DcpResponse.

If a close stream occurred we would add Buffer::bytes to the
FlowControl ack size and then also if the DcpResponse was also
processed the DcpResponse size is also added - a double count.

This is fixed by ensuring Buffer::bytes only accounts for the
items owned by the Buffer::messages queue.

An existing test was already exercising this scenario, but lacked
the correct EXPECTS to catch - now fixed.

Change-Id: Ib1b359d77ffe10bc2390906b187d35ac20e6a39e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166431
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>