MB-49626: getDroppedCollectionStats -> getDroppedCollectionItemCount
Currently we return a full set of PersistedStats from
getDroppedCollectionStats. We only really need the itemCount though
and in a future change when we move this to MagmaDbStats we won't
move the other stats.

Change the API to only return the item count rather than the full
set of stats.

Change-Id: I8e3eeee9c6b6d9fafcbd1233e973e71e490a5c72
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166521
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>