MB-47300: Log open() errors at caller
Some error conditions at open() are expected and benign in some cases,
eg vbucket file-not-found at the very first start-up.

To avoid irrelevant, misleading and excessive logging in such cases,
here we remove error logging from the inner openSpecificDBFile() and
let the caller decide when it's the case to log an error.

Change-Id: I3f9d676a02172480cace860cbea6b905dbc3a39e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166478
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>