MB-49782: [1/2] Avoid unnecessary move construction of ScopeTimer1
The default move constructor for ScopeTimer1 may lead to unexpected
`listener.stop()` calls.

The moved-out-of instance will call stop() on the contained listener
when destroyed, but so will the newly created instance on its
moved/copied version of the listener.

E.g., with SpanStopwatch this would lead to errant traces being
recorded by the moved-out-of instance, if the move has not been elided.

This patch simply avoids unnecessary move construction of the form:

 ScopeTimer1<Foo> timer({arg1, arg2});

There are no current uses of ScopeTimer1 which genuinely rely on move
construction or assignment.

Change-Id: I6c2d8dd489a1aa9b914752ec66ab93bf5276fbe5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166513
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>