MB-49626: Test different state in FailCompactKVStoreCall
Currently we test the item count in FailCompactKVStoreCall to
determine if we took the correct code paths in compactDb for magma.
In a future change when I move dropped collections to MagmaDbStats
we will not be able to test the item count here as it will be
incorrect (as we are testing in a hacky way). Test the dropped
collections disk state is correct (which is what we should really
be testing anyway) instead.

Change-Id: I6d29f3c2ed3eb6790985443a172dc092b4a86d5c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166520
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>