MB-49746: Reschedule CompactTask after a failure
For any internally requested compaction, reschedule after a failure.
If a compaction is exclusively an external request, no reschedule
occurs.

Note that 1 compaction task may end up serving both an internal and
external request - in that case we still reschedule on failure as
we cannot expect the external request to retry.

Change-Id: I2fbba7cbf767eaf0c5a678e49f0790f5d0fedebc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166759
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>