MB-49767: Fix and enhance logging in EPBucket::flushVBucket_UNLOCKED
Update the log messages in EPBucket::flushVBucket_UNLOCKED to log the
fact there from EPBucket::flushVBucket_UNLOCKED not
EPBucket::flushVBucket(). Also enhance the log messages for the high
completed/prepare seqno exceptions for when a flush tries to break their
monotonic property. By adding extra information regarding the min and
max seqno's that are meant to be flushed to disk and what the proposed
vbucket state was going to be.

Change-Id: I72231a6ffa475e6b963d816dbf63ac3a0ce4bbb5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166622
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>