MB-47604: Support concurrent flush and compaction in Nexus
Currently NexusKVStore prevents concurrent flushing and
compaction because it is/was hard to reason about and make
comparisons between the two underlying KVStores if one can
change change while "underneath us" and if one changes while
we perform some operation.

The changes to allow implicit compactions to run with Nexus
dealt with most cases that we need to consider for things
changing "underneath us". This change adds some extra updates
to the purgeSeqno as we may move the purgeSenqo in other places.

This change also deals with allow flushing and compaction to
run at the same time (and potentially see inconsistent before/after
states as a result of that). We attempt to detect if a
flush/compaction has run concurrently and skip checks if so.

Concurrent flush and compaction will be disabled for now, but is
enabled for the NexusKVStore unit tests. It will be enabled in
the future when Nexus tests are more stable.

Change-Id: I44ff199d07910c02dc85b2e36149513487a06ac2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/163588
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>