MB-49461: Allow MagmaKVStore to return a file handle even if DiskFull
MB-49461 showed that warmup can fail in a situation that the disk is
full.

This is due to a small window where we're able to successfully
create a MagmaKVStore without needing to be opened it in a ReadOnly
mode. But then the disk becomes full before run
Warmup::loadCollectionStatsForShard(). While running
loadCollectionStatsForShard() we try and create a KVFileHandle using
MagmaKVStore::makeFileHandle().

This fails as we're unable to Sync Magma's in memory state and disk.
This would normally be fine if we had opened Magma in a read only mode
as we would just return a ptr to an in memory snapshot. We should also
do this if Magma::SyncKVStore() returns Status::DiskFull, as we still
want to allow us to warmup so that a client can read data from
memcached.

We will also be okay in the situation the user tries to write data to
memcached as any further writes should fail and thus trigger us to
never enable traffic after we're warmed up. Also the KVFileHandle
returned will only ever be used for reading.

Change-Id: Ie60a5b6ff0dea53256accb797b30db444a5e1dad
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/166935
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>