MB-47604: Add collectionsPurged compaction stat [1/3]
MagmaKVStore and CouchKVStore implement collections purging slightly
differently. Whilst both do it via compaction, MagmaKVStore compacts
only the range of the items changed, whilst CouchKVStore compacts
the entire keyspace. As such it's not necessarily valid to compare
the results and callback invocations of compactDB calls if we are
purging collections. To determine if a compaction is purging
collections, add a new CompactionStat for the number of collections
purged during a compaction. This will be used in a future change to
skip comparisons in NexusKVStore if necessary.

Change-Id: Id5f822bb52b82026307ee25341e1de2ab8d97f8b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167003
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>