MB-49786: Increase default number of AuxIO threads
BackfillManagers are one-per-DcpProducer, but are limited in
concurrency by the number of AuxIO threads.

With 8 or fewer cores, currently a single AuxIO thread will be created.
This single thread must serve all backfills for all producers for
replication and other components. This has been seen to be insufficient
(MB-48693).

In heavy DGM, backfills will largely be IO bound so an increased
number running concurrently is not likely to add significant CPU
contention.

After this change, AuxIO threads will by default scale with CPU count,
up to a maximum of 8.

Change-Id: Ic752fef329c2e099e98689a5682c0957a577c680
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167038
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>