MB-49349: [Eph] Only update purge seqno for time based purges
Currently when purging any deleted item for an ephemeral vBucket
we move the purge seqno. This is not necessary, we only need to move
the purge seqno when we purge tombstones after the metadata purge
interval. Deletes with replacements (stale) and deletes in dropped
collections do not need to move the purge seqno. For the former we
will always stream the replacement and for the latter we only need the
collection events to move the purge seqno. This prevents unnecessary
rollbacks.

Change-Id: Ifbe9d5dedc3529876d7a818f2cea7345efa3f285
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167242
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>