MB-46378: Use atomic version of NonNegativeCounter only where needed
Audit and update all uses of NonNegativeCounter to use the new
(explicitly) atomic AtomicNonNegativeCounter version. This is in
preparation for changing NonNegativeCounter to be non-atomic; as that
adds unnecessary cost if the atomicty is not needed.

This is particulary relevent on architectures such as AArch64 whose
default memory ordering is weaker than x86; and hence atomic
operations are more expensive relative to non-atomic ones.

Change-Id: If78fd2b25909e0783f3a760c07955de0f3970411
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167313
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>