MB-48777: Prevent race between shutting down warmup and changing state
Before this patch, a race existed between Warmup::transition() and
Warmup::Stop(). Where Warmup::Stop() could change the WarmupState() to
Done. Before another warmup thread tried to call
WarmupState::transition() to another state of warmup. This would crash
memcached. As the transition() method would now try and move Warmup out
of the Done state which is not legal. This would normally be fine but we
as we don't try to change state if we're in the done state. The problem
is that the state can change between the done state check and the call
to WarmupState::transition().

To fix this modify WarmupState::legalTransition() to the current state
and to state as arguments. Then refactor WarmupState::transition() to
perform the read and check if we're currently in the done state. If not
and the transition is legal then try and update the state by performing
a compare exchange to ensure the current state is what we expect it to
be. If not re-try unless the current state is Done.

Change-Id: Ie1f1d6a0e2d9808fb60396d6010b99625916cff0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167380
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>