MB-41323: Compaction throttle: snooze waiting tasks forever
When scheduling a compaction, if too many compaction tasks are already
running we put the new task to sleep instead of running it
immediately. When a running compaction task completes it will wake up
a sleeping task.

However, we only sleep the task for 60s. I assume the intent here is
that 60s is sufficiently long for one of the currently-running
compaction tasks to complete; however with large vBuckets and/or slow
disks this isn't necessarily the case and the sleeping task could wake
and be run before the current compaction task has finished. This would
result in the compaction throttle limit being exceeded.

Note that these compactions can be triggered by both ns_server (which
has it's own concurrent compaction limit so arguably less of an issue)
and triggered by collection dropping - which immediately creates one
compaction Task per-vBucket and hence could easily exceed the limit if compaction is slow.

Address this by changing the sleep to "forever" - and relying solely
on the currently running compaction task(s) finishing to wake the
sleeper up.

Change-Id: I4dc2f46dacc5050e80a44bfa8898e0827f98bcd6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167585
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>