MB-50061: Update vBucket docCount on logical insert
Rather than update the dropped stats doc count to adjust by the
correct amount we must instead treat the first mutation in a
collection as an insert if a doc with the same key was written to
the old generation of a collection. We cannot adjust the dropped
stats count as we race with a concurrent purge of the collection.

Change-Id: I2456834edc76650843b5324d89854da03ff3d9a1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167501
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>