ep_testsuite: Use std::string_view for test_MB-19687_fixed
test_MB-19687_fixed makes use of a large number of stat names which
are represented using std::string. These keys are never modified, so
we do not need a mutable type such as std::string for them. Simplify
to std::string_view, which reduces the runtime cost of this test,
which can be noticable under sanitizers.

Change-Id: I9884c00385328bec3b367887048b8c8e67329b28
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167615
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>