MB-49727: Expire by Compaction/Pager observes CM Quota
Currently compaction and HT ejection bypass the CM Quota, so new
expired docs are allowed to be queued into the CM regardless of the CM
memory state. This patch prevents that, in particular:

Compaction: The compaction_exp_mem_threshold config param was used
  for avoiding that expirations by compaction take the full Bucket
  Quota. In this pacth we turn the BucketQuota check into a CMQuota
  check. The param becomes unsed, it will be removed in a follow-up.

Pager: No check of any kind on the system memory state. So this patch
  just introduces the CMQuota check.

Change-Id: I1a01f6a110fa9bef0a5916872209e7aa69f22673
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167644
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>