MB-50146: Don't read value from disk for CommitSyncWrite at backfill
Currently when we backfill if we find a CommitSyncWrite in the HT
then we force the CacheCallback to return to fetch the full item
for the prepare seqno.

Since 6.5.0 we have always sent CommitSyncWrites as Mutations rather
than Commits so we don't need the prepare seqno. As such, we are
unnecessarily fetching the Commit from disk.

In 7.1.0 and newer we don't even store Commits as CommitSyncWrites
on disk anymore, we store them instead as Mutations so we aren't
even gaining any more information by fetching the item.

Change-Id: I32f6b1fc1b50698d1c38ec88c76d21394b3fb2c4
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167821
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>