MB-35297: Remove complete from the DCPBackfill::cancel path
DCPBackfill::cancel is invoked when shouldCancel returns true.

shouldCancel only returns true if:

* ActiveStream is in state dead
* ActiveStream has gone (the std::weak_ptr cannot be locked)

In either of those cases the complete(bool) function (implemented at
the BySeqno/ByID level) should not be called - why?

1) If the ActiveStream has gone, complete() will fail to lock the
   weak_ptr.
2) If the ActieStream can be locked, but is in state dead, there's
   no action to take. It is pointless for example to call
   completeBackfill on a dead stream.

This commit simplifies the backfill code based on the above by removing
the needless use of complete from the cancel path - changing the function
from:

   complete(bool cancel)

to

   complete()

Change-Id: I01e786ada63888f46ac5eac01fd683b60d6f6717
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173320
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>