MB-35297: Use getVbucketId consistently in backfill classes
The Vbid is stored in the DCPBackfill so we still know the vbid
if the streamPtr could not be "locked". This leads to the backfill
sub-classes obtaining the Vbid from different sources.

This commit changes the backfill sub-classes so that they only read
from one place (to keep the code consistent) - getVbucketId is always
available irrespective of the state of streamPtr.

Change-Id: I0b05214453cf754712a65706e8aaf1de94b55a8c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/167987
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>