MB-35297: Return finished when create/scan are complete
When the DCPBackfill::run function returns

* backfill_success - the task will be scheduled for a future invocation
* backfill_finished - the task is deleted
* backfill_snooze -  the task will be scheduled for a future invocation
  after some time

When create/scan call complete they should not return backfill_success,
there is no need for a future call, instead return backfill_finished.

When DCPBackfillDisk sees backfill_finished returned it will now place
itself in State::done, removing the need for the subclasses to drive
the state to done (which is what complete would do).

With that change, compete is removed from the DCPBackfillDisk virtual
interface and is now implemented in the subclasses. This sub-class
version take a reference to the ActiveStream removing the need to
lock and handle failure.

This commit requires changes to tests because DCPBackfillDisk tasks
now require one less invocation to be done.

DCPBackfillMemoryBuffered is updated as well to keep the set of
backfill tasks consistent in how they are executed.

Change-Id: I55359e9286d111b4a7ce7d83e9a7c8a9a2aa0897
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173322
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>