MB-50124: Add possible stat keys to help text in mcstat
Add an optional argument to --help=statkey to dump the
available stat keys. It is also possible to just print info
of a specific stat by --help=<name of stat group>

In order to make sure we don't regress by adding new keys which
won't be listed in the help make sure that the stats parsing
on the server use the same list of legal stat groups.

Change-Id: If521719f3385a4051662ffb32bc5961d234a4bc4
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168143
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>