MB-49191: Use Warmup::isComplete() for ep_engine ep_warmup_thread
We have two versions of the ep_warmup_thread stat, one living in the
warmup stats groupa dn the other in the low cardinality stats in
ep_engine.cc. The latter currently checks the wrong condition as
it was not updated along with the former.

Create a new function in Warmup to return the correct state of the
stat rather than rely on checking the correct condition for both
usages.

Change-Id: I17b1b7e17eae3c7f319c0b5e1ef38faf936b62fe
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168293
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>