Refactor: use std::string as parameter for engine::init
It is what we use in the core when passing down to the
engines, and by doing so we don't need to deal with the
potential nullptr case in the engine (as it would never
be that).

This allows us to simplify the code in EPE::intialize
and return immediately if the configuration is empty (at
least the dbname must be present). And now we know that
we'll _always_ print out the configuration as part
of creating the bucket.

Ideally we should have passed on std::string_view, but
the engines currently use config_parser to parse the
data and it expects a null-terminated string.

Change-Id: I0e41f9b95aa6faec10a3de1c3d3518428353a830
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168366
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>