MB-50286: Handle pre-7.1 CommitSyncWrite metadata
Prior to MB-48033 (Neo), CommittedSyncWrites were stored on-disk as
queue_op::commit_sync_write and V3 metadata. However this was
optimized in MB-48033 to store a queue_op::mutation and V1 metadata
(as V1 is smaller).

However, we did not fully handle reading existing V3 commit_sync_write
from disk - for example during an offline upgrade. As a consequence
attempting to read such items from disk (for example during warmup)
results in a failure to make an Item:

     CRITICAL (b1) WarmupBackfillTask::run(): caught exception while running backfill - aborting warmup: makeItemFromDocInfo: Invalid queue_op:mutation

This is because while we allow the V3::commit_sync_write to be decoded
from disk in MetaData::MetaDataV3, we transparently convert the Commit
operation to mutation:

    queue_op getDurabilityOp() const {
        switch (operation) {
        ...
        case Operation::Commit:
            return queue_op::mutation;

Logically this is fine, however that falls foul of the logic in
makeItemFromDocInfo() which attempts to extract durability-related
information from V3 metadata, and doesn't exist to encounter a
queue_op::mutation (as per the above exception).

Address by accepting queue_op::mutation as a valid V3 durability op.

Change-Id: I16d2bd4891e232585630ff97b1052c10541bbfba
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168378
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>