Cleanup: setConflictResolutionMode should throw invalid_argument
Its current usage is to update the conflict resolution mode to
whatever is stored in the configuration object, and the
configuration object already validated that the input value
is one of the legal values (if it isn't parsing of the
configuration would fail and the error would be reported
there).

Keeping the extra logic around trying to check the return value
of the function and optionally log and fail (which would never
happen) just adds extra complexity and make people reading
the function think that it is a potential error case.

Change-Id: I052758d451c05c91c3a2fdf89b0fa815a98ee80d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168422
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>