Reduce cost of including probabilistic_counter.h
probabilistic_counter.h is included by hash_table.h (as HashTable
includes a ProbabilisticCounter member), and hash_table.h is itself
included by many other files.

probabilistic_counter.h is relatively expensive to include, as while
it is very small it includes <random> which costs ~135ms.

Move the function which actually requires <random> into it's own .cc
file. This reduces the cost of including probabilistic_counter.h - so
much so that probabilistic_counter.h is no longer listed in the top
headers, so using a header which includes it as proxy to measure the
improvement.

before:

    19130 ms: kv_engine/engines/ep/src/hash_table.h (included 77 times, avg 248 ms)

after:

    7755 ms: kv_engine/engines/ep/src/hash_table.h (included 77 times, avg 100 ms)

Change-Id: Ic6098e90ed1ee2e86cda4310754b14cf8c87de0a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168587
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>