MB-50333: Stream snapStart from Disk Checkpoint
Currently when we stream a Disk Checkpoint we take the seqno of the
first item in the Disk checkpoint and use that as the snapshot start
seqno. There exists a sanity check in the processing of aborts on the
replica that validates that the prepare seqno of an abort is within
the current snapshot if a prepare for it exists with a different seqno.
This sanity check can fail in the case where we receive the following
sequence of items.

1) Receive DCP disk snapshot with start seqno 1 end seqno 1
2) Receive DCP mutation at seqno 1
3) Receive DCP disk snapshot with start seqno 3 end seqno 3
4) Receive DCP abort at seqno 3 (seqno 2 was a de-duped prepare)

Fix this issue by sending the snapStart of the Disk checkpoint rather
than the seqno of the first item in the Disk checkpoint as the snapshot
start seqno (a snapshot marker with start seqno 2 end seqno 3)

In a future change we will make a similar change for memory snapshots,
but this is done separately as it affects non-replication DCP
consumers.

Change-Id: I6b5942ca506a2b320501031cc4b60df0e0dfb0cc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168617
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>