MB-50256: 3/n Update GetAllVbSeqs privilege checking for ReadSeqno
GetAllVBSeqs can operate with MetaRead or ReadSeqno depending on the
input.

Primarily the updated functionality is as follows.

When the command targets the bucket (no collection encoded) the command
requires ReadSeqno or MetaRead as bucket privilege.

When the command explictly targets a collection the command requires
MetaRead.

Note this commit also tidies up the input validation, if the client
has not enabled collections and they request a collection, fail the
command. This removes the previous ability to ask for a collection but
be given the wrong answer.

Change-Id: Ia3cf954ef4a36108070a4a4dc519708c0f52ecf6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168754
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>