MB-50400: Get _local/scope not _local/scopes
Problem:
MB-50400 observed a crash when upgrading from 7.0.0 to 7.1.0 for a
bucket that use scopes other than the default collection.

Cause:
This was caused a regression when modifying code so that it could be
used with Magma, specifically the add ding the const
LocalDocKey::openScopes, which was added with the value
"_local/scopes/open". This meant that after upgrade KVStore would look
under the wrong local key when trying to fined the vbucket's scope data.

Fix:
Change LocalDocKey::openScopes back to "_local/scope/open"

Change-Id: If8326938c2724400263402c755f35f860a3c52ce
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168766
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>