MB-50033: Include cert names found in CA root
Change the unit tests to use:

               <ROOT-CA>
               /       \
      <client cert>    <server cert>

The GO client will only send the client certificate if
it is belongs to the chain of CAs reported by the
server.

Previously we returned only the names found in the
certificate chain configured in memcached, but that
would (in this configuration) only be the server certificate
so that the go client would not send the certificate back.

This patch change that to also send the names of the
certificates stored in the CA file.

Change-Id: I68cf18bf0ce030f8dfddaabbede3248c388b7e78
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168842
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>