MB-50333: Set active Checkpoint snapStart to lastBySeqno + 1
Checkpoint snapStart is currently set to lastBySeqno when creating
checkpoints on active vBuckets. This means that the snapshot start
is set to the seqno of the last item in the previous checkpoint.
To set the snapshot start of the SnapshotMarkers that we send when
we cross checkpoint boundaries we first need to correct the snapshot
start to set it to the correct value (the seqno of the first item
in the snapshot).

Change-Id: Ibb68376f23026d8cfa2691313306aa93ed1c4f59
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168931
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>