MB-50383: End DCP legacy stream if default coll highSeqno is not visible
Problem:
MB-50383 demonstrated if a vbucket had no items in the default
collection due to following steps, we would fail the DCP stream instead
of returning a DcpStreamEnd with the flag set to status OK.

1. Create a second collection
2. Write a key to default collection and flush it
3. Delete the key in the default collection and flush
4. Write a key to second collection and flush
5. Perform compaction on the vbucket, this will purge the tombstone
   created at step 3., if enough time has passed for the purging
   threshold or compaction was performed with drop_deletes=true.

We will fail the stream as the high seqno vbucket is in the second
collection, which means we will allow the dropping of tombstone of step
3. This would then cause us to fail the backfill and set the stream to
dead due to KVStore::getBySeqno() return cb::engine_errc::no_such_key as
the tombstone has been purged.

Fix:
Instead of failing back fill in
DCPBackfillBySeqnoDisk::markLegacyDiskSnapshot() due to
KVStore::getBySeqno() returning cb::engine_errc::no_such_key, we should
instead continue to in the markLegacyDiskSnapshot() method and perform a
backfill to try and find the highest valid tombstone or alive document
that can be streamed to the client.

Also in the situation that the scan covers the full range of the DCP
stream i.e. the stream's streanEndSeqno == backfillEndSeqno. Then check
check if the endSeqno is greater than the maxVisableSeqno for the
default collection. As if this is the case we'll need to send a
DcpStreamEnd to the client. To do this we'll reduce the current
streamEnd to the maxVisableSeqno of the default collection that we found
when performing the scan of the vbucket on disk.

Change-Id: I70312f2f9beed1ec06d1b959150e73e8caa16192
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169030
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>