MB-49512: Remove magma checkpoint creation during shutdown
Doing a Sync(true) results in the Magma instance first waiting
for all flushes in the background queues to clear and then
creation of a checkpoint. It results in the first magma instance
being shutdown waiting on flushes of all other magma instances
since they share the threadpool. This causes the first magma
instance to take very long to shutdown(5+ min in this bug) followed
by the others taking under a few sec.

I have made changes in magma::Close to cancel queued background jobs.
Once that is done we internally create checkpoints.

Change-Id: Ib2270e6fccd8cba7e878149aeb23f6e2d17a921a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/169103
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>